+++ catscripts/Entertainment/Cartoons/Futurama/Futurama.sh
@@ -1,14 +1,24 @@
 #!/bin/bash
 
-egrep -i 'Phil(|l)ip(| )(J|)(\.|)(| )Fry|Leela|Bender(| )(|Bending)(| )Rodríguez|Planetary(| )Express|Zoidberg|Prof(\.||essor)(| )Farnsworth|Futurama' newpages.txt | egrep -iv 'Futurama' >> Futurama.txt
+KEYWORDS_FUTURAMA="Phil(|l)ip(| )(J|)(\.|)(| )Fry|Leela|Bender(| )(|Bending)(| )Rodríguez|Planetary(| )Express|Zoidberg|Prof(\.||essor)(| )Farnsworth|Futurama"
 
-FUTURAMA=$(stat --print=%s Futurama.txt)
-
-if [ $FUTURAMA -ne 0 ];
+if [ "$1" == "" ];
 then
-  export CATFILE="Futurama.txt"
-  export CATNAME="Futurama"
-  $CATEGORIZE
-fi
 
-rm Futurama.txt
+  debug_start "Futurama"
+
+  FUTURAMA=$(egrep -i "$KEYWORDS_FUTURAMA" newpages.txt)
+
+  if [ "$FUTURAMA" != "" ];
+  then
+    printf "%s" "$FUTURAMA" > Futurama.txt
+    export CATFILE="Futurama.txt"
+    export CATNAME="Futurama"
+    $CATEGORIZE
+    rm Futurama.txt
+    unset FUTURAMA
+  fi
+
+  debug_end "Futurama"
+
+fi
