+++ catscripts/Entertainment/Cartoons/Looney_Tunes/LooneyTunes.sh
@@ -2,15 +2,23 @@
 
 KEYWORDS_LOONEYTUNES="Bugs(| )Bunny|What(|')s(| )up(|,)Doc|Elmer(| )Fudd|Porky(| )Pig|Marvin(| )(|the)(| )Martian|Barnyard(| )Dawg|Goofy(| )Gophers|Foghorn(| )Leghorn|Speedy(| )Gonzales|Daffy(| )Duck|Sylvester(| )the(| )Cat|\bTweety\b|Wile(| )E(|\.)(| )Coyote|Pep(e|é)(| )Le(| )Pew"
 
-egrep -i "$KEYWORDS_LOONEYTUNES" newpages.txt >> LooneyTunes.txt
-
-LOONEY=$(stat --print=%s LooneyTunes.txt)
-
-if [ $LOONEY -ne 0 ];
+if [ "$1" == "" ];
 then
-  export CATFILE="LooneyTunes.txt"
-  export CATNAME="Looney Tunes"
-  $CATEGORIZE
-fi
 
-rm LooneyTunes.txt
+  debug_start "Looney Tunes"
+
+  LOONEY=$(egrep -i "$KEYWORDS_LOONEYTUNES" newpages.txt)
+
+  if [ "$LOONEY" != "" ];
+  then
+    printf "%s" "$LOONEY" > LooneyTunes.txt
+    export CATFILE="LooneyTunes.txt"
+    export CATNAME="Looney Tunes"
+    $CATEGORIZE
+    rm LooneyTunes.txt
+    unset LOONEY
+  fi
+
+  debug_end "Looney Tunes"
+
+fi
